requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
langgraph==0.5.1
langchain==0.3.26
langchain-openai==0.3.27
langchain-community==0.3.27
langchain-core==0.3.68
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.1
typing-extensions==4.14.1
pandas==2.1.4
requests==2.31.0 